refactor(page-editor): split oversized files to satisfy LOC cap
Extract React decorator views, block renderers, inline-format helpers,
and the chip-resolver into sibling modules so every node + walker file
stays under the 150 effective-line cap. Pure mechanical split — no
behavior changes; the test helper move quiets a separate
consistent-function-scoping rule.
