+++ src/features/page-editor/nodes/SignatureBlockNode.tsx
@@ -1,7 +1,5 @@
-import { useState } from "react";
 import {
   $applyNodeReplacement,
-  $getNodeByKey,
   DecoratorNode,
   type DOMConversionMap,
   type DOMExportOutput,
@@ -11,19 +9,15 @@ import {
   type SerializedLexicalNode,
   type Spread,
 } from "lexical";
-import { useLexicalComposerContext } from "@lexical/react/LexicalComposerContext";
-import { interpolate } from "@/features/templates/interpolate";
-import { EditPropModal } from "../EditPropModal";
-import { useLetterVars } from "../letterRenderContext";
-import { LETTER_VARIABLES } from "../letterVariables";
+import { SignatureBlockView } from "./SignatureBlockView";
 
 export type SerializedSignatureBlockNode = Spread<
   { closing: string; signatory: string },
   SerializedLexicalNode
 >;
 
-const DEFAULT_CLOSING = "With gratitude,";
-const DEFAULT_SIGNATORY = "The Bishopric";
+export const DEFAULT_CLOSING = "With gratitude,";
+export const DEFAULT_SIGNATORY = "The Bishopric";
 
 /** Block-level decorator that renders the closing-of-letter signature
  *  line: a soft phrase ("With gratitude,"), a short walnut underline,
@@ -110,63 +104,13 @@ export class SignatureBlockNode extends DecoratorNode<React.ReactElement> {
 
   decorate(): React.ReactElement {
     return (
-      <SignatureView nodeKey={this.__key} closing={this.__closing} signatory={this.__signatory} />
+      <SignatureBlockView
+        nodeKey={this.__key}
+        closing={this.__closing}
+        signatory={this.__signatory}
+      />
     );
   }
-}
-
-function SignatureView({
-  nodeKey,
-  closing,
-  signatory,
-}: {
-  nodeKey: NodeKey;
-  closing: string;
-  signatory: string;
-}) {
-  const [editor] = useLexicalComposerContext();
-  const vars = useLetterVars();
-  const [editing, setEditing] = useState<"closing" | "signatory" | null>(null);
-
-  function save(next: string) {
-    editor.update(() => {
-      const node = $getNodeByKey(nodeKey);
-      if (!(node instanceof SignatureBlockNode)) return;
-      if (editing === "closing") node.setClosing(next.trim() || DEFAULT_CLOSING);
-      else if (editing === "signatory") node.setSignatory(next.trim() || DEFAULT_SIGNATORY);
-    });
-    setEditing(null);
-  }
-
-  return (
-    <>
-      <div contentEditable={false} className="select-none mt-7 mb-2 [&_.sig-rule]:border-walnut-3">
-        <button
-          type="button"
-          onClick={() => setEditing("closing")}
-          className="block font-serif italic text-[16px] text-walnut mb-2 text-left hover:underline focus:outline-none"
-        >
-          {interpolate(closing, vars)}
-        </button>
-        <div className="sig-rule border-b border-walnut-3 w-[260px] mb-1.5" />
-        <button
-          type="button"
-          onClick={() => setEditing("signatory")}
-          className="block font-mono text-[10px] tracking-[0.18em] uppercase text-walnut-3 text-left hover:underline focus:outline-none"
-        >
-          {interpolate(signatory, vars)}
-        </button>
-      </div>
-      <EditPropModal
-        open={editing !== null}
-        title={editing === "signatory" ? "Edit signed-by line" : "Edit closing phrase"}
-        initial={editing === "signatory" ? signatory : closing}
-        variables={LETTER_VARIABLES}
-        onSave={save}
-        onCancel={() => setEditing(null)}
-      />
-    </>
-  );
 }
 
 export function $createSignatureBlockNode(
